requirements.txt
aiosmtplib==4.0.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
brotlipy==0.7.0
bump2version==1.0.1
bumpversion==0.6.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
extruct==0.18.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
Flask==3.1.0
greenlet==3.2.2
h11==0.16.0
html5lib==1.1
html_text==0.7.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
id==1.5.0
idna==3.10
incremental==24.7.2
isort==6.0.1
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jstyleson==0.0.2
keyring==25.6.0
Levenshtein==0.27.1
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mf2py==2.0.1
more-itertools==10.7.0
motor==3.7.0
msgpack==1.1.0
nh3==0.2.21
packaging==25.0
parsel==1.10.0
passlib==1.7.4
pdoc3==0.11.6
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyRdfa3==3.6.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
RapidFuzz==3.13.0
rdflib==7.1.4
readme_renderer==44.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rich-toolkit==0.13.2
SecretStorage==3.3.3
sentry-sdk==2.28.0
service-identity==24.2.0
setuptools==80.4.0
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.1
tldextract==5.3.0
twine==6.1.0
typer==0.15.2
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.4
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
zope.interface==7.2
